fix type

+++ database/migrations/2016_03_23_192827_add_more_fields_to_users_table.php
@@ -16,7 +16,7 @@ class AddMoreFieldsToUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->integer('default_game_type_id')->unsigned();
+            $table->tinyInteger('default_game_type_id')->unsigned();
             $table->foreign('default_game_type_id')->references('id')->on('game_types')->onDelete('cascade');
             $table->boolean('default_game_private');
         });
